% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContrastCoefficients.R
\name{ContrastCoefficients}
\alias{ContrastCoefficients}
\title{Extract all binary contrast coefficients of interest from a regression model}
\usage{
ContrastCoefficients(
  mod,
  Data,
  pred,
  ixterm = NULL,
  robust = T,
  HCtype = "HC0",
  usedf = F,
  savecontrastmodels = F
)
}
\arguments{
\item{mod}{This is a model object of class \code{\link[stats]{lm}},
\code{\link[stats]{glm}}, \code{negbin} (see \code{\link[MASS]{glm.nb}}),
\code{\link[logistf]{logistf}}, or a \code{list}
or \code{\link[mice]{mira}} class list of models, where in each model is a
\code{lm}, \code{glm}, or \code{negbin} object. If \code{mod} is a
\code{list} or \code{mira} object, estimates will be pooled using the
\code{\link[mice]{pool}} function.}

\item{Data}{This is a data frame used to run the model \code{mod}. This
should be of class \code{data.frame} unless \code{mod} is of class
\code{list} or \code{mira}, in which case \code{Data} should be a list of
datasets of class \code{list} or \code{\link[mice]{mids}}.}

\item{pred}{This is a character value defining the column name of the
predictor of interest in the model \code{mod} and the data frame \code{Data}.}

\item{ixterm}{This is an optional character value defining the column name
of the interacting variable included in the model \code{mod} and present in
the data frame \code{Data}. This defaults to \code{NULL}, meaning that no
interacting variable contrasts will be output. It is necessary for there to
be an interaction with the variable \code{ixterm} in the model \code{mod} for
the function to work. Also, the \code{pred} variable should be on the left
side of the interaction term while \code{ixterm} is on the right side (i.e.,
\code{pred:ixterm} and not \code{ixterm:pred}).}

\item{robust}{If \code{TRUE}, this calculates and returns robust CIs and
p-values based on the \code{\link[sandwich]{vcovHC}} function instead of the
default CIs and p-values. This defaults to \code{TRUE}, and is only applied
if \code{mod} is of class \code{'lm'} or \code{'glm'}.}

\item{HCtype}{This is the formula used for the \code{\link[sandwich]{vcovHC}}
function calculation of robust CIs and p-values. Defaults to \code{"HC0"}.}

\item{usedf}{If \code{FALSE}, this sets \code{\link[lmtest]{coefci}} to
calculate z-score-based robust confidence intervals if \code{robust} is
\code{TRUE}. If \code{TRUE}, t test-based robust confidence intervals are
instead calculated based on the residual degrees of freedom in each
regression if \code{robust} is \code{TRUE}. This defaults to \code{FALSE},
which is the default for the \code{df} argument in
\code{\link[lmtest]{coefci}}.}

\item{savecontrastmodels}{If \code{TRUE}, this will output the additional
iterated models (beyond the original model \code{mod}) as a list. This
defaults to \code{FALSE}.}
}
\value{
If \code{savecontrastmodels} is \code{FALSE},
\code{ContrastCoefficients} will simply return a matrix of the coefficients,
their error and confidence interval estimates, and their p-values. Otherwise,
\code{ContrastCoefficients} will return a list of the following items:
\item{Coefficients}{This is the matrix of coefficients and related values
returned by this function regardless.}
\item{ContrastModels}{This is the list of iterated model objects. If
\code{mod} was a \code{list} or \code{mira} object, this will be a list of
lists of model objects.}
}
\description{
\code{ContrastCoefficients} is a function designed to extract all the binary
combinations of contrasts for a categorical (factor) predictor of interest
from a linear model. The typical output of linear model only returns the
coefficients for categorical variables in the context of a single referent
level. For example, if a predictor variable of interest is a 4-level
categorical treatment variable with factor levels
\code{c("placebo", "drug1", "drug2", "drug3")}, a linear model will only
return the "drug1 - placebo", "drug2 - placebo", and "drug3 - placebo" binary
contrast coefficients. \code{ContrastCoefficients} will return all binary
contrast coefficients for all possible referent levels, meaning the
aforementioned coefficients will be returned, as well the coefficients for
"drug2 - drug1", "drug 3 - drug 1", and "drug 3 - drug 2" contrasts.
Furthermore, \code{ContrastCoefficients} can also provide all contrasts for
an interaction term as well. This works for when both the predictor of
interest and the interacting variable are categorical or when either the
predictor of interest or the interacting variable are categorical. If only a
continuous predictor of interest or both a continuous predictor and
interacting variable are specified, the function will simply return the
coefficients of interest.
}
\details{
The way this function obtains these additional coefficients is by iteratively
updating the referent level for the predictor of interest and, if specified,
the interacting variable and rerunning the linear models. Other methods may
instead try to calculate these coefficients from the initial model
coefficients, but the iterative method has the benefit of uniformly applying
the same error estimation procedures for each coefficient. Some helpful
internal functions used in this functions include
\code{DrewDayRFunctions:::getcombos} to find all n combinations within one or
between two categorical variables and
\code{DrewDayRFunctions:::getmodelcoefs}, which extracts all model
coefficients within a given model (i.e., without providing all contrast
combinations) for the specified predictor and/or interacting variable of
interest for the specified model.

This function is currently set up to support \code{\link[stats]{lm}},
\code{\link[stats]{glm}}, \code{\link[MASS]{glm.nb}},
and \code{\link[logistf]{logistf}} (i.e., Firth's
logistic regression) model objects, as well as a \code{list}
or \code{\link[mice]{mira}} class list of models produced by the
multiple imputation function \code{\link[mice]{mice}}, wherein each model is
either a \code{lm}, \code{glm}, or \code{glm.nb} model.

\code{ContrastCoefficients} returns model coefficients for all unique binary
combinations of categorical predictor variables when there is no interaction
term specified in the model and the predictor variable of interest is
categorical. When there is an interaction term, \code{ContrastCoefficients}
returns all marginal coefficients for all unique binary contrast
coefficients at each referent level of the interaction term are reported as
"marginal coefficients".

As an example, we will define Model 1 with a
continuous predictor "age" and a two-level interaction term "sex" and Model 2
with a 3-level categorical predictor "treatment" and the same two-level
categorical interaction term "sex". If \code{allmarginalcontrasts = TRUE},
then the Model 1 coefficients will include age given sex = male (
"age|sex = male"), "age|sex = female", the interacting variable coefficient
"sex (female - male)", and the interaction term coefficient
"age:sex (female - male)". For Model 2, the coefficients would include
\enumerate{
\item all binary treatment contrasts at each sex
referent level (i.e., "treatment (2 - 1)|sex = male",
"treatment (2 - 1)|sex = female", "treatment (3 - 1)|sex = male",
"treatment (3 - 1)|sex = female", etc.), 2. the sex
coefficient "sex (female - male)", and 3. all unique binary contrast
interaction term coefficients (i.e., "treatment (2 - 1):sex (female - male)",
"treatment (3 - 1):sex (female - male)", &
"treatment (3 - 2):sex (female - male)").
}

Note that this function does not support greater than two-way interactions
(e.g., three-way interactions or higher), and it also does not support
interactions that do not include the main effects of both the predictor of
interest and the interacting variable (i.e., formulae with
\code{pred * ixterm} or \code{pred + ixterm + pred:ixterm} are okay, but
not those with \code{pred + pred:ixterm} or \code{ixterm + pred:ixterm}).
}
\examples{
# Note that these examples with the infert dataset do not reflect any real 
# hypotheses and are instead just to demonstrate the function with example 
# continuous and categorical variables.

data("infert")
set.seed(5)
infert$fac4 <- as.factor(sample(1:4, nrow(infert), replace = T))

#A categorical predictor without any interaction terms
mod.1 <- glm(case ~ fac4 + education + age, data = infert, family = binomial)
ContrastCoefficients(mod.1, infert, "fac4")

#A categorical predictor with a categorical interaction term
mod.1 <- glm(case ~ fac4*education + age, data = infert, family = binomial)
ContrastCoefficients(mod.1, infert, "fac4", "education")

#A categorical predictor with a continuous interaction term
mod.1 <- glm(case ~ fac4*age + education, data = infert, family = binomial)
ContrastCoefficients(mod.1, infert, "fac4", "age")

#A continuous predictor with a categorical interaction term
mod.1 <- glm(case ~ age*fac4 + education, data = infert, family = binomial)
ContrastCoefficients(mod.1, infert, "age", "fac4")

#Missing data imputation with mice & a categorical by categorical interaction
infert$education_missing <- infert$education
set.seed(3)
infert[sample(1:nrow(infert), 50), "education_missing"] <- NA
infert.mice <- mice::mice(infert[, c(
"education_missing", "case", "fac4", "age")], maxit = 3, m = 3, seed = 20, 
printFlag = F)
# maxit and m should be higher, but this is just a fast-running example
mod.1 <- with(data = infert.mice, 
exp = glm(case ~ fac4*education_missing + age, family = binomial))
ContrastCoefficients(mod.1, infert.mice, "fac4", "education_missing")

}
