% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionCoefPlot.R
\name{InteractionCoefPlot}
\alias{InteractionCoefPlot}
\title{Visualize continuous by continuous interactions}
\usage{
InteractionCoefPlot(
  model,
  data,
  pred,
  ixterm,
  multiplier = 1,
  coeftransform = NULL,
  predname = NULL,
  ixname = NULL,
  outname = NULL,
  title = NULL,
  fillcolor = NULL,
  autotitle = F,
  addpvallab = F,
  labsize = 4,
  lengthout = 50,
  shadebysig = F,
  otherix = F,
  robust = T,
  HC = "HC0",
  logistic = F,
  xlab10exp = F
)
}
\arguments{
\item{model}{A model of class \code{\link[stats]{lm}} or \code{\link[stats]{glm}}.}

\item{data}{The data frame used to generate \code{model}.}

\item{ixterm}{The column name of the interacting variable in the data frame
\code{data}.}

\item{multiplier}{A multiplier for the coefficients of interest (e.g., an IQR
of the predictor of interest). Defaults to 1.}

\item{coeftransform}{An optional transformation function for the coefficients
(e.g., \code{tenfoldperc} if the predictor of interest is on a log10 scale
where \code{tenfoldperc <- function(x) ((10^x) - 1) * 100)}. Defaults to
\code{NULL}.}

\item{predname}{An optional new name for the variable \code{pred}. Defaults to
\code{NULL}.}

\item{ixname}{An optional new name for \code{ixterm}. Defaults to \code{NULL}.}

\item{outname}{An optional new name for the outcome variable. Defaults to
\code{NULL}.}

\item{title}{An optional plot title. Defaults to \code{NULL}.}

\item{fillcolor}{The color for the shading of the 95\% CI region. Defaults to
\code{"grey50"}.}

\item{autotitle}{An option to automatically generate a plot title. Defaults to
\code{FALSE}.}

\item{addpvallab}{An option to add a label to the plot that shows the
interaction term p-value. Defaults to \code{FALSE}.}

\item{labsize}{The label text size if \code{addpvallab} is \code{TRUE}. Defaults
to 4.}

\item{lengthout}{The number of levels of \code{ixterm} generated with the
\code{length.out} argument for the \code{\link[base]{seq}} function, defaults
to 50.}

\item{shadebysig}{An option to shade the CI region areas that don't overlap
the null a darker tint of \code{fillcolor}. Defaults to \code{FALSE}.}

\item{otherix}{An option to include a second plot in which \code{pred} and
\code{ixterm} are flipped. Defaults to \code{FALSE}.}

\item{robust}{This is an option to use robust sandwich errors for the CIs
(see \code{\link[sandwich]{vcovHC}} for details on HC estimators). Defaults to
\code{TRUE}. This should be set to \code{TRUE} if model is a robust Poisson
model for RR estimation.}

\item{HC}{This is a character value defining the robust estimator used if
\code{robust = TRUE} (see \code{\link[sandwich]{vcovHC}} for details on HC
estimators). This defaults to \code{"HC0"}.}

\item{logistic}{An option for if model is a logistic \code{\link[stats]{glm}}.
Defaults to \code{FALSE}. This option sets the null line to 1, exponentiates
all the coefficients to bring them to an odds ratio scale, and changes the y
label to "OR" instead of "Coefficient".}

\item{xlab10exp}{An option to change the x-axis labels to be of the form
"10^x" (see \code{\link[scales]{label_math}}). This is useful if your
predictor of interest (or also the \code{ixterm}
variable if \code{otherix} is \code{TRUE}) is on a logarithmic scale.}
}
\value{
\code{InteractionCoefPlot} returns a list of:
\item{MainGGplot}{The interaction plot}
\item{OtherGGplot}{The other interaction plot if \code{otherix = TRUE}}
\item{MainMatrix}{The matrix of values used to generate the main plot}
\item{OtherMatrix}{The matrix of values used to generate the other plot if
\code{otherix = TRUE}}
}
\description{
\code{InteractionCoefPlot} produces a plot of how the coefficient between
one variable in a bivariate continuous interaction and the outcome changes
over levels of the other continuous variable. This is a useful visualization
of bivariate interactions between two continuous variables.
}
\details{
Plot axis labels, fonts, etc. can be changed with standard ggplot options.
For example, if the model is a robust Poisson GLM, you can set
\code{logistic = TRUE} when running this function and saving it into an
object that in this example I will call \code{plotlist} and then change the
"OR" label to "RR" by running \code{plotlist$MainGGplot + ylab("Main Variable 
RR")} or something similar.
}
\examples{

set.seed(1)
exdat <- as.data.frame(mvtnorm::rmvnorm(500, mean=rep(0,5)))
names(exdat) <- paste0("X", 1:ncol(exdat))
exdat$X1X2 <- exdat$X1 * exdat$X2
exdat$yhat <- c(-1 + as.matrix(exdat) \%*\% c(0.5, 1.5, 0, -2, 2, 1))
set.seed(2)
exdat$y <- exdat$yhat + rnorm(500,0,5)
exlm <- lm(y ~ X1 * X2 + X3 + X4 + X5, exdat)
explot <- InteractionCoefPlot(exlm, exdat, "X1", "X2", addpvallab=T, 
shadebysig=T, robust=F, otherix=T)
if(require(cowplot)){
cowplot::plot_grid(explot$MainGGplot, explot$OtherGGplot, ncol=1)}

}
