% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericDataSummary.R
\name{GenericDataSummary}
\alias{GenericDataSummary}
\title{Generate an html R Markdown document briefly summarizing a dataset}
\usage{
GenericDataSummary(
  filename = NULL,
  directory = getwd(),
  Data,
  Dict = NULL,
  idvar = NULL
)
}
\arguments{
\item{filename}{A file name and path for the .html R Markdown document to be
generated. Defaults to "DataSummary_X.html" where X is the system date and
time when the document is saved. You can specify the ".html" suffix here or
omit it, in which case it will be added for you.}

\item{directory}{The folder in which you'd like to save the html document.
This defaults to your current working directory.}

\item{Data}{A data frame to be summarized.}

\item{Dict}{An optional data frame for the data dictionary to be included for
reference as a table in the Markdown document. This defaults to NULL.}

\item{idvar}{An optional single character value indicating the column name of
a column of IDs in Data. If specified, a short statement on the number of
unique IDs and counts of rows per ID will be added to the introduction of
the document. This defaults to NULL.}
}
\value{
\code{GenericDataSummary} saves an html document to the specified
filename and directory.
}
\description{
\code{GenericDataSummary} creates a Markdown document for any dataset that
separately summarizes continuous variables in that dataset with a summary
table, boxplots, and histograms, as well as provides frequency counts for
all categorical data. There is also an option to include a data dictionary
table in this simple summary document. Uses for this function include
quickly inspecting a dataset for aberrant values and getting a sense for
which variables are included.
}
\details{
This Markdown document makes use of \link[DT]{datatable} from the DT package
for the data dictionary and summary statistics tables, which can be
sorted and searched. Categorical variable counts are summarized in a
datatable with a column showing percentages with a bar plot in the
background.
}
\examples{
# Make a data dictionary for the built-in iris dataset
irisdd <- data.frame(Variable = names(iris), 
Description = gsub("\\\\.", " ", names(iris)), Levels = "")
irisdd[irisdd$Variable == "Species", "Levels"] <- 
paste(unique(iris$Species), collapse = ", ")

# create the summary document
GenericDataSummary("NewIrisSummary", getwd(), iris, irisdd)

}
