% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.medcurve.R
\name{plot.medcurve}
\alias{plot.medcurve}
\title{Plot the results of the function MediationCurve}
\usage{
\method{plot}{medcurve}(
  med.results,
  plot.est = "ACME",
  addci = T,
  smoothplot = T,
  addpts = F,
  addptci = F,
  plotrug = F,
  rugvalues = NULL,
  xsplineshape = -0.5
)
}
\arguments{
\item{med.results}{This is an output from the function \code{MediationCurve}
with the class \code{'medcurve'}.}

\item{plot.est}{This is the estimate of interest to be plotted. Options
include \code{"ACME"} for the average causal mediation effect, \code{"ADE"}
for the average direct effect, \code{"Total"} for the total effect, and
\code{"Proportion"} for the proportion of the effect mediated (i.e.,
ACME/Total Effect). This defaults to \code{"ACME"}.}

\item{addci}{This is a logical value specifying whether 95\% confidence
intervals should be wrapped around the mean estimate curve or line.
This defaults to \code{TRUE}.}

\item{smoothplot}{This is a logical value specifying whether the main line
and 95\% confidence interval ribbon (if \code{addci = TRUE}) connecting the
point estimates should be a smoothed x-spline fit
(\code{\link[ggalt]{stat_xspline}}) if \code{TRUE} or a series
of straight lines connecting the point estimates if \code{FALSE}. The
x-spline fit forces the curve to go through each observed point. This
defaults to \code{TRUE}.}

\item{addpts}{This is a logical value specifying whether to add points for
the point estimates of the estimate of interest. This defaults to
\code{FALSE}.}

\item{addptci}{This is a logical value specifying whether to add error bars
for the 95\% confidence intervals of the point estimates of the estimate of
interest. This defaults to \code{FALSE}.}

\item{plotrug}{This is a logical value specifying whether to add a rug plot
along the x-axis showing observed treatment values from the data used to
fit the models. If \code{TRUE}, a vector of treatment values must be
provided for the argument \code{rugvalues}. This defaults to \code{FALSE}.}

\item{rugvalues}{This is a numeric vector of the treatment values to be
included in the rug plot if \code{plotrug = TRUE}. This defaults to
\code{NULL}. If \code{plotrug = TRUE} and \code{rugvalues} is NULL, an error
will be returned.}

\item{xsplineshape}{This is a value between -1 and 1 controlling the
smoothing of the x-spline (see \code{\link[ggalt]{stat_xspline}}) smooth.
The x-spline approximates a Catmull-Rom spline at -1, has sharp corners at 0,
and approximates a cubic b-spline at 1. This defaults to -0.5, which has the
x-spline smoothly curving directly through all points.}
}
\value{
\code{plot.medcurve} returns the specified plot of the class
\code{'ggplot'}.
}
\description{
This function plots one of the estimates of interest from the output of the
\code{MediationCurve} function as either a smoothed curve or a series of
straight lines connecting the point estimates. Point x-axis values are
plotted at the mean of each treatment 'bin'.
}
\examples{
\dontrun{
#Smoothed ACME fit
plot(medres)

#Non-smoothed ADE fit with rug plot
plot(medres, "ADE", smoothplot = F, plotrug = T, rugvalues = meddata$treat)

#Smoothed Total Effect fit with point estimate points and CIs
plot(medres, "Total", addpts = T, addptci = T)
}

}
