% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.medcurve.R
\name{plot.medcurve}
\alias{plot.medcurve}
\title{Plot the results of the function MediationCurve}
\usage{
\method{plot}{medcurve}(
  med.results,
  plot.est = "ACME",
  addci = T,
  smoothplot = T,
  addpts = T,
  addptci = T,
  plotrug = F,
  rugvalues = NULL,
  xsplineshape = -0.5,
  ribbon = F,
  addcurve = T,
  addstep = F
)
}
\arguments{
\item{med.results}{This is an output from the function \code{MediationCurve}
with the class \code{'medcurve'}.}

\item{plot.est}{This is the estimate of interest to be plotted. Options
include \code{"ACME"} for the average causal mediation effect, \code{"ADE"}
for the average direct effect, \code{"Total"} for the total effect, and
\code{"Proportion"} for the proportion of the effect mediated (i.e.,
ACME/Total Effect). This defaults to \code{"ACME"}.}

\item{addci}{This is a logical value specifying whether 95\% confidence
intervals should be wrapped around the mean estimate curve or line.
This defaults to \code{TRUE}.}

\item{smoothplot}{This is a logical value specifying whether the main line
and 95\% confidence interval ribbon (if \code{addci = TRUE}) connecting the
point estimates should be a smoothed x-spline fit
(\code{\link[ggalt]{stat_xspline}}) if \code{TRUE} or a series
of straight lines connecting the point estimates if \code{FALSE}. The
x-spline fit forces the curve to go through each observed point. This
defaults to \code{TRUE}.}

\item{addpts}{This is a logical value specifying whether to add points for
the point estimates of the estimate of interest. This defaults to
\code{FALSE}.}

\item{addptci}{This is a logical value specifying whether to add error bars
for the 95\% confidence intervals of the point estimates of the estimate of
interest. This defaults to \code{FALSE}.}

\item{plotrug}{This is a logical value specifying whether to add a rug plot
along the x-axis showing observed treatment values from the data used to
fit the models. If \code{TRUE}, a vector of treatment values must be
provided for the argument \code{rugvalues}. This defaults to \code{FALSE}.}

\item{rugvalues}{This is a numeric vector of the treatment values to be
included in the rug plot if \code{plotrug = TRUE}. This defaults to
\code{NULL}. If \code{plotrug = TRUE} and \code{rugvalues} is NULL, an error
will be returned.}

\item{xsplineshape}{This is a value between -1 and 1 controlling the
smoothing of the x-spline (see \code{\link[ggalt]{stat_xspline}}) smooth.
The x-spline approximates a Catmull-Rom spline at -1, has sharp corners at 0,
and approximates a cubic b-spline at 1. This defaults to -0.5, which has the
x-spline smoothly curving directly through all points.}

\item{ribbon}{This is a logical value specifying whether the confidence
interval curves (if \code{addci = TRUE}) will be plotted as curves (if
\code{FALSE}) or as a filled-in area specified by \code{geom_ribbon} (if
\code{TRUE}). This values defaults to \code{TRUE}. In some cases (e.g.,
sometimes with \code{plot.est = "Proportion"}), the confidence interval
curves can be sufficiently different lengths the x-values of the CI curve
boundaries will be different lengths and not coincide enough to merge to form
the dataset used by geom_ribbon to fill in the ribbon area. If that's the
case, a warning will print and the argument \code{'ribbon'} will be ignored.
This defaults to \code{FALSE}.}

\item{addcurve}{This is a logical value specifying whether a curve linking
the mediation estimate values (and the CI values if \code{addci = TRUE})
should be drawn. One may want to plot just the points and CIs without linking
them with a curve line, in which case one should set this argument to
\code{FALSE} and specify \code{addpts = TRUE} and \code{addptci = TRUE}. This
defaults to \code{TRUE}.}

\item{addstep}{This is a logical value specifying whether the differences
between successive pairs of control and treated values if control values
vary in the \code{medcurve} object should be plotted as a step function. In
this visualization, the control and treated values form the boundaries of
each step, and the y-axis value of each step is the estimate of interest for
that set of control and treated values. This defaults to \code{FALSE}.}
}
\value{
\code{plot.medcurve} returns the specified plot of the class
\code{'ggplot'}.
}
\description{
This function plots one of the estimates of interest from the output of the
\code{MediationCurve} function as either a smoothed curve or a series of
straight lines connecting the point estimates. Point x-axis values are
plotted at the mean of each treatment 'bin'.
}
\examples{
\dontrun{
#Smoothed ACME fit
plot.medcurve(medres)

#Non-smoothed ADE fit with rug plot
plot.medcurve(medres, "ADE", smoothplot = F, plotrug = T, 
rugvalues = meddata$treat)

#Smoothed Total Effect fit with point estimate points and CIs
plot.medcurve(medres, "Total", addpts = T, addptci = T)
}

}
